Move version-related tests
